requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
dnspython>=2.4.0
google-generativeai>=0.8.0
